feat: Making commitlint compatible with dependabot
Dependabot commit subjects generate scoped commit types (scoping
type to `deps`).  We want this behavior, and do not want build to
fail on autogenerated dependabot commits due to our rule that prevents
scope in commit type.
The second problem this addresses is increasing max body line to 100
characters, allowing dependabot to generate documentation linking to
security docs for the update